+++ docs/Range.js
@@ -1,0 +1,12 @@
+/**
+ * @category Types
+ * @summary An object that combines two dates to represent the range.
+ *
+ * @description
+ * An object that combines two dates to represent the range.
+ *
+ * @typedef {Object} Range
+ * @param {Date|String|Number} start - the start of the range
+ * @param {Date|String|Number} end - the end of the range
+ */
+var Range = {}