+++ src/is_within_range/test.js
@@ -8,8 +8,7 @@ describe('isWithinRange', function () {
   it('returns true if the given date in within the given range', function () {
     var result = isWithinRange(
       new Date(2014, 9 /* Oct */, 31),
-      new Date(2014, 8 /* Sep */, 1),
-      new Date(2014, 11 /* Dec */, 31)
+      {start: new Date(2014, 8 /* Sep */, 1), end: new Date(2014, 11 /* Dec */, 31)}
     )
     assert(result === true)
   })
@@ -17,8 +16,7 @@ describe('isWithinRange', function () {
   it('returns true if the given date has same time as the left boundary of the range', function () {
     var result = isWithinRange(
       new Date(2014, 8 /* Sep */, 1),
-      new Date(2014, 8 /* Sep */, 1),
-      new Date(2014, 11 /* Dec */, 31)
+      {start: new Date(2014, 8 /* Sep */, 1), end: new Date(2014, 11 /* Dec */, 31)}
     )
     assert(result === true)
   })
@@ -26,8 +24,7 @@ describe('isWithinRange', function () {
   it('returns true if the given date has same time as the right boundary of the range', function () {
     var result = isWithinRange(
       new Date(2014, 11 /* Dec */, 31),
-      new Date(2014, 8 /* Sep */, 1),
-      new Date(2014, 11 /* Dec */, 31)
+      {start: new Date(2014, 8 /* Sep */, 1), end: new Date(2014, 11 /* Dec */, 31)}
     )
     assert(result === true)
   })
@@ -35,8 +32,7 @@ describe('isWithinRange', function () {
   it('returns true if the given date and the both boundaries are the same', function () {
     var result = isWithinRange(
       new Date(2014, 11 /* Dec */, 31),
-      new Date(2014, 11 /* Dec */, 31),
-      new Date(2014, 11 /* Dec */, 31)
+      {start: new Date(2014, 11 /* Dec */, 31), end: new Date(2014, 11 /* Dec */, 31)}
     )
     assert(result === true)
   })
@@ -44,8 +40,7 @@ describe('isWithinRange', function () {
   it('returns false if the given date is outside of the range', function () {
     var result = isWithinRange(
       new Date(2014, 1 /* Feb */, 11),
-      new Date(2014, 8 /* Sep */, 1),
-      new Date(2014, 11 /* Dec */, 31)
+      {start: new Date(2014, 8 /* Sep */, 1), end: new Date(2014, 11 /* Dec */, 31)}
     )
     assert(result === false)
   })
@@ -53,8 +48,7 @@ describe('isWithinRange', function () {
   it('accepts a string', function () {
     var result = isWithinRange(
       new Date(2014, 9 /* Oct */, 31).toISOString(),
-      new Date(2014, 8 /* Sep */, 1).toISOString(),
-      new Date(2014, 11 /* Dec */, 31).toISOString()
+      {start: new Date(2014, 8 /* Sep */, 1).toISOString(), end: new Date(2014, 11 /* Dec */, 31).toISOString()}
     )
     assert(result === true)
   })
@@ -62,8 +56,7 @@ describe('isWithinRange', function () {
   it('accepts a timestamp', function () {
     var result = isWithinRange(
       new Date(2014, 9 /* Oct */, 31).getTime(),
-      new Date(2014, 8 /* Sep */, 1).getTime(),
-      new Date(2014, 11 /* Dec */, 31).getTime()
+      {start: new Date(2014, 8 /* Sep */, 1).getTime(), end: new Date(2014, 11 /* Dec */, 31).getTime()}
     )
     assert(result === true)
   })
@@ -72,8 +65,7 @@ describe('isWithinRange', function () {
     var block = isWithinRange.bind(
       null,
       new Date(2014, 9 /* Oct */, 31),
-      new Date(2014, 11 /* Dec */, 31),
-      new Date(2014, 8 /* Sep */, 1)
+      {start: new Date(2014, 11 /* Dec */, 31), end: new Date(2014, 8 /* Sep */, 1)}
     )
     assert.throws(block)
   })